test-client/requirements.txt
python-socketio==5.11.1
aiohttp==3.9.3
colorama==0.4.6
websockets==12.0 